move data population utilities into a utils module to avoid duplicating methods for migrations
