+++ src/scripts/populate.js
@@ -3,27 +3,22 @@
  * database with static fixtures at `cve-services/datadump/pre-population`
  */
 
-const express = require('express')
 const config = require('config')
-const mongoose = require('mongoose')
-const uuid = require('uuid')
+const express = require('express')
 const app = express()
+const mongoose = require('mongoose')
+
+const dataUtils = require('../utils/data')
+const errors = require('../utils/error')
 const logger = require('../middleware/logger')
 const CveIdRange = require('../model/cve-id-range')
 const CveId = require('../model/cve-id')
 const Cve = require('../model/cve')
 const Org = require('../model/org')
 const User = require('../model/user')
-const argon2 = require('argon2')
-const cryptoRandomString = require('crypto-random-string')
-const CONSTANTS = require('../constants')
-const utils = require('../utils/utils')
-const fs = require('fs')
-const JSONStream = require('JSONStream')
-const util = require('util')
-const errors = require('../utils/error')
+
 const error = new errors.IDRError()
-const apiKeyFile = 'user-secret.txt'
+
 require('dotenv').config() // This enables us to read from the .env file.
 
 const populateTheseCollections = {
@@ -118,40 +113,39 @@ db.once('open', async () => {
       names.push(collection.name)
     })
 
-    // TODO: need to populate org, user, cve-id in that order given UUID links
-
-    if (!names.includes('Cve-Id-Range') && !names.includes('Cve-Id') && !names.includes('Cve') && !names.includes('Org') && !names.includes('User')) {
+    if (!names.includes('Cve-Id-Range') && !names.includes('Cve-Id') && !names.includes('Cve') &&
+        !names.includes('Org') && !names.includes('User')) {
       // Org
-      await populateCollection(
+      await dataUtils.populateCollection(
         './datadump/pre-population/orgs.json',
-        Org, newOrgTransform
+        Org, dataUtils.newOrgTransform
       )
 
       // User, depends on Org
-      const hash = await preprocessUserSecrets()
-      await populateCollection(
+      const hash = await dataUtils.preprocessUserSecrets()
+      await dataUtils.populateCollection(
         './datadump/pre-population/users.json',
-        User, newUserTransform, hash
+        User, dataUtils.newUserTransform, hash
       )
 
       let populatePromises = []
 
       // CVE ID Range
-      populatePromises.push(populateCollection(
+      populatePromises.push(dataUtils.populateCollection(
         './datadump/pre-population/cve-ids-range.json',
         CveIdRange
       ))
 
       // CVE
-      populatePromises.push(populateCollection(
+      populatePromises.push(dataUtils.populateCollection(
         './datadump/pre-population/cves.json',
         Cve
       ))
 
       // CVE ID, depends on User and Org
-      populatePromises.push(populateCollection(
+      populatePromises.push(dataUtils.populateCollection(
         './datadump/pre-population/cve-ids.json',
-        CveId, newCveIdTransform
+        CveId, dataUtils.newCveIdTransform
       ))
 
       // don't close database connection until all remaining populate
@@ -167,114 +161,3 @@ db.once('open', async () => {
     }
   }
 })
-
-async function preprocessUserSecrets () {
-  const color = require('kleur')
-
-  if (process.env.NODE_ENV === 'development') {
-    const secretKey = process.env.LOCAL_KEY
-    const hash = await argon2.hash(secretKey)
-
-    // provide secret to user in development
-    console.log(color.bold().black().bgWhite(
-      'Use the following API secret for all users:') + ' ' +
-      color.bold().black().italic().bgGreen(secretKey)
-    )
-
-    return hash
-  } else {
-    // delete file for user secrets if one already exists
-    fs.unlink(apiKeyFile, err => {
-      if (err && err.code !== 'ENOENT') {
-        logger.error(error.fileDeleteError(err))
-        mongoose.connection.close()
-      }
-    })
-
-    console.log(
-      color.bold().black().bgWhite('The users\' API secret can be found in:') + ' ' +
-      color.bold().black().italic().bgGreen(apiKeyFile)
-    )
-  }
-}
-
-async function newOrgTransform (org) {
-  org.UUID = uuid.v4()
-  org.inUse = false
-  return org
-}
-
-async function newUserTransform (user, hash) {
-  const tmpOrgUUID = await utils.getOrgUUID(user.cna_short_name)
-  user.org_UUID = tmpOrgUUID
-  user.UUID = uuid.v4()
-  user.authority = { active_roles: [] }
-
-  // shared secret key in development environments
-  if (process.env.NODE_ENV === 'development') {
-    user.secret = hash
-  } else {
-    const randomKey = cryptoRandomString({ length: CONSTANTS.CRYPTO_RANDOM_STRING_LENGTH })
-    user.secret = await argon2.hash(randomKey)
-
-    // write each user's API key to file
-    // necessary when standing up any new shared instance of the system
-    const payload = { username: user.username, secret: randomKey }
-    fs.writeFile(apiKeyFile, JSON.stringify(payload) + '\n', { flag: 'a' }, (err) => {
-      if (err) {
-        logger.error(error.fileWriteError(err))
-        mongoose.connection.close()
-      }
-    })
-  }
-
-  return user
-}
-
-async function newCveIdTransform (cveId) {
-  const tmpRequestingCnaUUID = await utils.getOrgUUID(cveId.requested_by.cna)
-  const tmpOwningCnaUUID = await utils.getOrgUUID(cveId.owning_cna)
-  const tmpUserUUID = await utils.getUserUUID(cveId.requested_by.user, tmpRequestingCnaUUID)
-  cveId.requested_by.cna = tmpRequestingCnaUUID
-  cveId.owning_cna = tmpOwningCnaUUID
-  cveId.requested_by.user = tmpUserUUID
-  return cveId
-}
-
-/* populates any collection given the file path for an input JSON,
- * the data model for the collection, a function that transforms the
- * data, and a hash that only gets use in the User collection
- */ 
-function populateCollection (filePath, dataModel, dataTransform = async function (x) {return x}, hash) {
-  const dataName = dataModel.collection.collectionName
-  logger.info(`Populating ${dataName} collection...`)
-  return new Promise(function (resolve, reject) {
-    // let batchData = []
-    let promises = []
-    fs.createReadStream(filePath, {encoding: 'utf8'})
-      .pipe(JSONStream.parse('*'))
-      .on('data', async function (data) {
-        this.pause()
-        data = await dataTransform(data, hash)
-        promises.push(dataModel.insertMany(data))
-        this.resume()
-        promisesShifter(promises)
-      })
-      .on('close', function () {
-        Promise.all(promises).then(function () {
-          logger.info(`${dataName} populated!`)
-          resolve()
-        })
-      })
-      .on('error', reject)
-  })
-}
-
-function promisesShifter (promises) {
-  if (promises) {
-    while (!util.inspect(promises[0]).includes('pending')) {
-      promises.shift()
-      if (promises.length == 0) { break }
-    }
-  }
-}
